+++ src/test/unit/components/Creator/GridOptions.test.tsx
@@ -1,10 +1,10 @@
-import { renderComponent, renderWithContext } from "../testUtils"
+import { renderComponent, renderWithContext } from "../../../testUtils"
 import { fireEvent, screen } from '@testing-library/react'
-import { GridOptions } from "../../components/creator/Editor/SceneEdition/GridOptions/GridOptions"
-import { LocalStorage } from "../../localStorage"
-import { SceneMap, SerializedChallenge, defaultChallenge, defaultScene } from "../../components/serializedChallenge"
-import { ACTOR, EMPTY, OBSTACLE } from "../../components/creator/Editor/SceneEdition/mapUtils"
-import { CreatorContextProvider } from "../../components/creator/Editor/CreatorContext"
+import { GridOptions } from "../../../../components/creator/Editor/SceneEdition/GridOptions/GridOptions"
+import { LocalStorage } from "../../../../localStorage"
+import { SceneMap, SerializedChallenge, defaultChallenge, defaultScene } from "../../../../components/serializedChallenge"
+import { ACTOR, EMPTY, OBSTACLE } from "../../../../components/creator/Editor/SceneEdition/mapUtils"
+import { CreatorContextProvider } from "../../../../components/creator/Editor/CreatorContext"
 
 describe('Scene grid', () => {
