+++ src/test/unit/components/Creator/SceneEdition.test.tsx
@@ -1,8 +1,8 @@
 import { fireEvent, screen } from '@testing-library/react'
-import { defaultChallenge, SerializedChallenge } from "../../components/serializedChallenge"
-import { SceneEdition } from '../../components/creator/Editor/SceneEdition/SceneEdition';
-import { LocalStorage } from '../../localStorage';
-import { renderWithContext } from '../testUtils';
+import { defaultChallenge, SerializedChallenge } from "../../../../components/serializedChallenge"
+import { SceneEdition } from '../../../../components/creator/Editor/SceneEdition/SceneEdition';
+import { LocalStorage } from '../../../../localStorage';
+import { renderWithContext } from '../../../testUtils';
 
 describe('Scene Edition', () => {
     afterEach(() => {
