+++ src/test/unit/components/Creator/StatementEdition.test.tsx
@@ -1,8 +1,8 @@
 import { fireEvent, screen } from '@testing-library/react'
-import { defaultChallenge } from "../../components/serializedChallenge"
-import { StatementEdition } from '../../components/creator/Editor/ChallengeDetailsEdition/StatementEdition';
-import { LocalStorage } from '../../localStorage';
-import { renderComponent } from '../testUtils';
+import { defaultChallenge } from "../../../../components/serializedChallenge"
+import { StatementEdition } from '../../../../components/creator/Editor/ChallengeDetailsEdition/StatementEdition';
+import { LocalStorage } from '../../../../localStorage';
+import { renderComponent } from '../../../testUtils';
 
 describe('Statement Edition', () => {
     afterEach(() => {
