+++ src/test/unit/components/MainRoutesComponents.test.tsx
@@ -1,14 +1,14 @@
-import { About } from "../components/about/About"
-import { BookView } from "../components/book/BookView"
-import { ChallengeById, ChallengeByName } from "../components/EmberChallengeView"
-import { CreatorEditor } from "../components/creator/Editor/Editor"
-import { ActorSelection } from "../components/creator/ActorSelection/ActorSelection"
-import { Home } from "../components/home/Home"
-import { ImportedChallengeView } from "../components/ImportedChallengeView"
-import { ChangePassword, PasswordRecovery } from "../components/users/passwordRecovery/PasswordRecovery"
-import { Register } from "../components/users/register/Register"
-import { expectToThrow, renderComponent } from "./testUtils"
-import { validChallenge } from "./serializedChallenge.test"
+import { About } from "../../../components/about/About"
+import { BookView } from "../../../components/book/BookView"
+import { ChallengeById, ChallengeByName } from "../../../components/EmberChallengeView"
+import { CreatorEditor } from "../../../components/creator/Editor/Editor"
+import { ActorSelection } from "../../../components/creator/ActorSelection/ActorSelection"
+import { Home } from "../../../components/home/Home"
+import { ImportedChallengeView } from "../../../components/ImportedChallengeView"
+import { ChangePassword, PasswordRecovery } from "../../../components/users/passwordRecovery/PasswordRecovery"
+import { Register } from "../../../components/users/register/Register"
+import { expectToThrow, renderComponent } from "../../testUtils"
+import { validChallenge } from "../serializedChallenge.test"
 test('Renders home without errors', async () => {
   expect(() => renderComponent(<Home />)).not.toThrow()
 })
